+++ package.json
@@ -28,6 +28,7 @@
     "express-session": "^1.15.6",
     "mongoose": "^5.0.11",
     "mongoose-file": "0.0.2",
+    "octopartjs": "^1.0.5",
     "passport": "^0.4.0",
     "passport-local": "^1.0.0",
     "passport-local-mongoose": "^5.0.0",
