+++ routes/index.js
@@ -2,6 +2,7 @@ const path = require("path");
 const router = require("express").Router();
 const apiRoutes = require("./api");
 
+
 // API Routes
 router.use("/api", apiRoutes);
